requirements.txt
# SmartOrder Engine - Зависимости Python

# ===========================================
# Web Framework
# ===========================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# ===========================================
# Database
# ===========================================
psycopg2-binary==2.9.9

# ===========================================
# HTTP Client
# ===========================================
requests==2.31.0
urllib3==2.1.0

# ===========================================
# Redis
# ===========================================
redis==5.0.1

# ===========================================
# Task Scheduler
# ===========================================
APScheduler==3.10.4

# ===========================================
# Telegram Bot
# ===========================================
python-telegram-bot==20.7

# ===========================================
# AI / OpenAI
# ===========================================
openai==1.12.0

# ===========================================
# Fuzzy Matching
# ===========================================
rapidfuzz==3.6.1

# ===========================================
# PDF Generation
# ===========================================
reportlab==4.0.7

# ===========================================
# Excel Export
# ===========================================
openpyxl==3.1.2

# ===========================================
# Utilities
# ===========================================
python-dotenv==1.0.0
python-json-logger==2.0.7

# ===========================================
# Rate Limiting
# ===========================================
slowapi==0.1.9

# ===========================================
# HTTP Client (для тестирования API - опционально)
# ===========================================
httpx==0.25.2